Add support for absolute terminal paths and fix a bug
Add support for passing in a absolute path of the terminal location and
fix a bug with an early return.

Previously the for loop at the end of the searchForBinary function would
return once it found a program that exists but  is not executable. Now
it continiues to check all the paths until it finds a executable program
and then returns.